Extract validators, utilities, and common code
